Add HTML generation and tests
This is a bit of a "draw the rest of the fucking owl" change, but I didn't want to commit extra broken code. There's a fair bit of test cleanup that needs to happen here, and some further testing with more complicated codebases, but it's working in some primitive cases, and I'm happy for that right now.