Handle conditionally executed chunks of function bodies
The issue was if there was a function with a condition or other body-like statement that was marked as covered by the coverprofile, tarp wouldn't override that. So you'd end up with really goofy looking reports where the start of the function would be yellow, but it would resume being either green or red when the coverprofile encountered a change. This commit makes it so that tarp only lets that behavior happen if we're not in the uncovered function, and the line wouldn't otherwise be marked red.